wip: fix status list json-ld verification and cleanup verifier code
